Update installation files to work
The largest change is moving from R 3.5.1 to R 4.1. This change is
required as the installation is failing on a dependent package limma.
An updated version of limma which is required to work on Apple Silicon
is only available for newer versions of R.

Minor changes to versions of tqdm, MarkupSafe, and certifi, for
dependency resolution issues.

Changed conda commands to require no interaction by adding -y flag.